# libcxx: Optimizations for uniform_int_distribution