I noticed that the implementation for `std::uniform_int_distribution` in libc++ often has worse performance than other implementations.. While it might be useful to change the implementation to one that rejects less samples, such a change would result in the output of the distribution changing, which could break programs dependent on the behavior. Instead, this PR attempts to optimize the existing implementation without changing the output of the distribution.

In libc++, `std::uniform_int_distribution` uses `__independent_bits_engine` to produce n-bit random numbers. The implementation of `__independent_bits_engine` is based on the specification for `std::independent_bits_engine`, except that the value for n is specified as a runtime value. However, the design of `std::independent_bits_engine` was optimized for the case where n is known at compile time. The algorithm used requires calculating a number of values, which `std::independent_bits_engine` can calculate at compile time, avoiding any runtime performance penalty. However, the use of this algorithm for `std::uniform_int_distribution` results in these values having to be calculated for every output generated, which ends up accounting for a significant portion of the algorithm's runtime.

While analyzing the definitions of the values that need to be generated and the algorithm in general, I identified two common cases where the algorithm could be simplified:

* When R is a power of two, y0 = R, and if it's used, y1 = R. This means that the algorithm will never reject any samples generated, allowing it to use a simplified implementation of `__eval` that limits branching and doesn't read the values for y0 or y1 (allowing the optimizer to potentially eliminate the variables entirely). Additionally, the check for the condition R - y0 <= floor(y0 / n) simplifies to 0 <= floor(R / n), which is always true, meaning we can skip this check entirely when generating the required values.
* When w <= m, we know that the result of evaluating the required values will always result in n = 1. This is because, when n = ceil(w / m) = 1, R - y0 <= floor(y0 / n) can be simplified to just R <= 2\*y0, which is always true. (You can see it's always true by noting that the definitions for y0 and m gives y0 >= 2^m and R < 2^(m+1)). With this, we can provide an optimized implementation for this case that skips the multiple divisions that are required for larger values of w.

Additionally, I noticed that pretty much every use of a bit shift in the original implementation was guarded by an if clause that ensured the shift value did not cause UB. However, these checks introduce complexity and branching that can impact performance. I was able to remove these if statements by carefully considering what all the possible shift values can be. Note, I'm working with the assumption that 0 < w <= numeric_limits<result_type>::digits, which appears to be a safe assumption to make about w given how `__independent_bits_engine` is used by `std::uniform_int_distribution`.

* For computing mask0, the method used is already safe. The definition of w0 ensures that 0 < w0 <= EDt, so the value for (EDt - w0) is in the range 0 <= (EDt - w0) < EDt, which is the valid range for the shift.
* For computing y0, the current method requires checking that w0 < EDt, since it's possible that w0 = EDt. This optimization handles this case by changing the evaluation of y0 to not use any shifts, and instead use the value of mask0. This prevents the possibility of UB from an invalid shift value by simply not performing any shifts.
* For computing mask1 and y1, we have that if n > n0, then w0 + 1 <= EDt. This means that this case can be handled with the same method used for mask0 and y0. For the case where n = n0, then the values for mask1 and y1 go unused, and so we don't need to calculate them anyways. To avoid uninitialized variable warnings, this change defaults to using mask1 = mask0 and y1 = y0.
* For the shifts in `__eval`, the case where n > 1 will never result in a shift value that results in UB. This is because the relation n0\*w0 + (n - n0)\*(w0 + 1) = w gives that w0 < w, and if n > n0, w0 + 1 < w. This means the only case that needs to be handled is when n = 1. This is done by manually unrolling the k = 0 iteration and removing the shift step from it. This makes the assumptions that n > 0 and n0 > 0, but this is always true given w > 0 and n - n0 < n respectively (the latter being given by the "mod n" in the definition of n0).

All together, these changes provide minor performance improvements for sampling large ranges and significant performance improvements for sampling small ranges. The only other thing to note with this is that it removes the specific (and optimized) case for when Rp = 0 (representing R = 2^WDt) that was present previously. However, the optimizations for power-of-two values for R apply to this case as well, and there doesn't appear to be much benefit to trying to further optimize this case. This change also reuses the overload for `__eval` previously used for the Rp = 0 special case for handling power-of-two values of R.