@LRFLEW 2025-05-15
I included two test files I made while validating this change as part of this PR.

The first is a test file that tests the output of `std::uniform_int_distribution` against lists of expected outputs. The lists were generated using LLVM 19.1.7, and are intended as a means of validating that the changes I made didn't change the output of the distribution. Since the C++ standard doesn't specify how distributions are supposed to generate values, this test does not validate standard conformance at all, and only tests if the output has changed relative to previous LLVM versions.

The second is a benchmark for `std::uniform_int_distribution`. It tests the distribution using minstd_rand0 and ranlux24_base (chosen for their speed, which minimizes the impact of the PRNG performance on the measured time), which tests both when R isn't and is a power of two respectively. Both engines are tested using three different sample ranges: 2^20, chosen to result in n = 1, 2^40, chosen to result in n = 2 and n0 = 2, and 2^41, chosen to result in n = 2 and n0 = 1. This gives good coverage of the different possible results based on my optimization efforts.

I ran this test on my machine (i7 8700K, WSL on Windows 10) with and without the changes made in this PR, took the median time of 16 repetitions, and compiled the result in the table below. This should give a good idea of the impact this change will have.

| Engine and Sample Size | Original Time | New Time | Performance |
| --- | --- | --- | --- |
| minstd_rand0, w = 20 | 39.3 ns | 13.9 ns | 2.83x |
| ranlux24_base, w = 20 | 47.9 ns | 23.0 ns | 2.08x |
| minstd_rand0, w = 40 | 45.3 ns | 38.8 ns | 1.17x |
| ranlux24_base, w = 40 | 61.3 nss | 46.7 ns | 1.31x |
| minstd_rand0, w = 41 | 45.8 ns | 40.3 ns | 1.14x |
| ranlux24_base, w = 41 | 61.2 ns | 48.7 ns | 1.26x |

I kept these test file additions in its own commit preceding the changes to the libc++ header. I did this because a) it makes it easier to perform the before-and-after tests, and b) I'm not sure if these test files should be merged. While the benchmark *may* be useful, the other test file likely isn't useful, since it doesn't test anything related to standard conformance or a known regression. I'll leave the tests as they are now for code review. However, let me know if you want me to squash the commits or remove any of the test files, and I'll do it.